fix the logic of finding current master in cluster
The origin logic is to find whether one of partners is master nor not. For most cases that none of the partners are masters. It is only one ref to find the right number.

The new logic is to find current cluster master by the endpoints in partners, and expose the right master to start entrypoint.